Move dropdown component to common folder (as reusable presentation component)
