+++ src/components/projecteditor/control/item/fileItem/directoryEntry.js
@@ -23,7 +23,7 @@ import {
     IconImportFile
 } from '../../../../icons';
 import style from '../../style.less';
-import { DropdownContainer } from '../../../../dropdown';
+import { DropdownContainer } from '../../../../common';
 import { FadeInComponent } from './fadeInComponent';
 import { ShowActions } from './showActions';
 import Tooltip from '../../../../tooltip';
